Move testcase_utils to _internal/common
